refactor(chat): extract surfaceToSpec to dedicated file with UIElement types
